More code sharing between search and field pages
