+++ frontends/mit-open/src/pages/FieldPage/FieldSearchFacetDisplay.tsx
@@ -9,7 +9,7 @@ import type {
 } from "@mitodl/course-search-utils"
 import { BOOLEAN_FACET_NAMES } from "@mitodl/course-search-utils"
 import { Skeleton, styled } from "ol-components"
-import { StyledDropdown } from "../SearchPage/SearchPage"
+import { StyledDropdown } from "@/page-components/SearchDisplay/SearchDisplay"
 
 export type KeyWithLabel = { key: string; label: string }
